+++ TrafficModel/VehicleType.cpp
@@ -4,12 +4,12 @@
 /// \brief Create a new VehicleType object.
 /// \param src The source node.
 /// \param dst The destination node.
-VehicleType::VehicleType(uint8_t src, uint8_t dst) {
+VehicleType::VehicleType(uint16_t src, uint16_t dst) {
   _trip[0] = src;
   _trip[1] = dst;
 }
-uint8_t VehicleType::getSource() const noexcept { return _trip[0]; }
-uint8_t VehicleType::getDestination() const noexcept { return _trip[1]; }
+uint16_t VehicleType::getSource() const noexcept { return _trip[0]; }
+uint16_t VehicleType::getDestination() const noexcept { return _trip[1]; }
 /// \brief Set the transition matrix.
 /// \param transMatrix The transition matrix.
 void VehicleType::setTransMatrix(std::vector<std::vector<double>> &matrix) {
@@ -17,7 +17,7 @@ void VehicleType::setTransMatrix(std::vector<std::vector<double>> &matrix) {
   if (n == 0)
     throw std::runtime_error("VehicleType::setTransMatrix: matrix is empty.\n");
   for(auto const& row : matrix) {
-	  	std::vector<double> temp;
+		std::vector<double> temp;
 		for(auto const& el : row) {
 			temp.push_back(el);
 		}
